Added initial HTML structure for portfolio website
This commit introduces the basic HTML structure for the portfolio website. Included are meta tags for charset and viewport, along with links to external CSS and Bootstrap framework. The page title is left blank for future customization. Additionally, a heading "Portfolio van Ayman Berkane" is added as the main content placeholder.